pyproject.toml
[project]
name = "words"
version = "0.0.0"
description = "Empty Django and Vue starter extracted from the shows harness."
authors = [{name = "Anton Samarchyan", email = "desecho@gmail.com"}]
license = {text = "MIT"}
readme = "README.rst"
requires-python = ">=3.13"

dependencies = [
    "Django>=5.2.4,<6",
    "django-cors-headers>=4.4.0,<5",
    "django-rest-registration>=0.9.0,<1",
    "djangorestframework>=3.13.1,<4",
    "djangorestframework-simplejwt>=5.5.1,<6",
    "django-stubs-ext>=5.2.2,<6",
    "mysqlclient>=2.2.0,<3",
    "gunicorn>=23.0.0,<24",
    "tzdata>=2025.2",
    "openpyxl>=3.1.5",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-django",
    "mypy",
    "django-stubs>=5.2.2,<6",
    "djangorestframework-stubs",
    "ruff",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/words"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
